--- docs/answers-react-components.filters.facetsprops.children.md
@@ -1,18 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProps](./answers-react-components.filters.facetsprops.md) &gt; [children](./answers-react-components.filters.facetsprops.children.md)
-
-## Filters.FacetsProps.children property
-
-A function which renders the Facets UI with the provided facets data.
-
-<b>Signature:</b>
-
-```typescript
-children?: (facets: DisplayableFacet[]) => ReactNode;
-```
-
-## Remarks
-
-It is intended to be used with the Filters subcomponents including Filters.FilterGroup, Filters.CollapsibleLabel, Filters.SearchInput, Filters.CheckboxOption, Filters.CollapsibleSection
-
