--- docs/answers-react-components.filters.facetsprops.classname.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProps](./answers-react-components.filters.facetsprops.md) &gt; [className](./answers-react-components.filters.facetsprops.classname.md)
-
-## Filters.FacetsProps.className property
-
-CSS class names applied to the component's container div.
-
-<b>Signature:</b>
-
-```typescript
-className?: string;
-```
