--- docs/answers-react-components.filters.facetsprops.md
@@ -1,22 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProps](./answers-react-components.filters.facetsprops.md)
-
-## Filters.FacetsProps interface
-
-Props for [Filters.Facets()](./answers-react-components.filters.facets.md)
-
-<b>Signature:</b>
-
-```typescript
-export interface FacetsProps 
-```
-
-## Properties
-
-|  Property | Type | Description |
-|  --- | --- | --- |
-|  [children?](./answers-react-components.filters.facetsprops.children.md) | (facets: DisplayableFacet\[\]) =&gt; ReactNode | <i>(Optional)</i> A function which renders the Facets UI with the provided facets data. |
-|  [className?](./answers-react-components.filters.facetsprops.classname.md) | string | <i>(Optional)</i> CSS class names applied to the component's container div. |
-|  [searchOnChange?](./answers-react-components.filters.facetsprops.searchonchange.md) | boolean | <i>(Optional)</i> Whether or not a search is automatically run when a filter is selected. Defaults to true. |
-
