--- docs/answers-react-components.filters.facetsprops.searchonchange.md
@@ -1,13 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProps](./answers-react-components.filters.facetsprops.md) &gt; [searchOnChange](./answers-react-components.filters.facetsprops.searchonchange.md)
-
-## Filters.FacetsProps.searchOnChange property
-
-Whether or not a search is automatically run when a filter is selected. Defaults to true.
-
-<b>Signature:</b>
-
-```typescript
-searchOnChange?: boolean;
-```
