+++ docs/answers-react-components.filters.facetsprovider.md
@@ -1,8 +1,8 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [Facets](./answers-react-components.filters.facets.md)
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProvider](./answers-react-components.filters.facetsprovider.md)
 
-## Filters.Facets() function
+## Filters.FacetsProvider() function
 
 The Facets component is a wrapper component around [Filters](./answers-react-components.filters.md) that updates facet options when a child filter is updated.
 
@@ -11,14 +11,14 @@ The representation of the facets is configured using a FACC (function as a child
 <b>Signature:</b>
 
 ```typescript
-export declare function Facets({ children, className, searchOnChange }: FacetsProps): JSX.Element;
+export declare function FacetsProvider({ children, className, searchOnChange }: FacetsProviderProps): JSX.Element;
 ```
 
 ## Parameters
 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
-|  { children, className, searchOnChange } | [FacetsProps](./answers-react-components.facetsprops.md) |  |
+|  { children, className, searchOnChange } | FacetsProviderProps |  |
 
 <b>Returns:</b>
