+++ docs/answers-react-components.filters.facetsproviderprops.children.md
@@ -1,0 +1,18 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProviderProps](./answers-react-components.filters.facetsproviderprops.md) &gt; [children](./answers-react-components.filters.facetsproviderprops.children.md)
+
+## Filters.FacetsProviderProps.children property
+
+A function which renders the Facets UI with the provided facets data.
+
+<b>Signature:</b>
+
+```typescript
+children?: (facets: DisplayableFacet[]) => ReactNode;
+```
+
+## Remarks
+
+It is intended to be used with the Filters subcomponents including Filters.FilterGroupProvider, Filters.CollapsibleLabel, Filters.SearchInput, Filters.CheckboxOption, Filters.CollapsibleSection
+
