+++ docs/answers-react-components.filters.facetsproviderprops.classname.md
@@ -1,0 +1,13 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProviderProps](./answers-react-components.filters.facetsproviderprops.md) &gt; [className](./answers-react-components.filters.facetsproviderprops.classname.md)
+
+## Filters.FacetsProviderProps.className property
+
+CSS class names applied to the component's container div.
+
+<b>Signature:</b>
+
+```typescript
+className?: string;
+```
