+++ docs/answers-react-components.filters.facetsproviderprops.md
@@ -1,0 +1,22 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FacetsProviderProps](./answers-react-components.filters.facetsproviderprops.md)
+
+## Filters.FacetsProviderProps interface
+
+Props for [Filters.FacetsProvider()](./answers-react-components.filters.facetsprovider.md)
+
+<b>Signature:</b>
+
+```typescript
+export interface FacetsProviderProps 
+```
+
+## Properties
+
+|  Property | Type | Description |
+|  --- | --- | --- |
+|  [children?](./answers-react-components.filters.facetsproviderprops.children.md) | (facets: DisplayableFacet\[\]) =&gt; ReactNode | <i>(Optional)</i> A function which renders the Facets UI with the provided facets data. |
+|  [className?](./answers-react-components.filters.facetsproviderprops.classname.md) | string | <i>(Optional)</i> CSS class names applied to the component's container div. |
+|  [searchOnChange?](./answers-react-components.filters.facetsproviderprops.searchonchange.md) | boolean | <i>(Optional)</i> Whether or not a search is automatically run when a filter is selected. Defaults to true. |
+
