--- docs/answers-react-components.filters.filtergroupprops.md
@@ -1,16 +1,0 @@
-<!-- Do not edit this file. It is automatically generated by API Documenter. -->
-
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FilterGroupProps](./answers-react-components.filters.filtergroupprops.md)
-
-## Filters.FilterGroupProps type
-
-Props for a [Filters.FilterGroup()](./answers-react-components.filters.filtergroup.md)<!-- -->.
-
-<b>Signature:</b>
-
-```typescript
-export declare type FilterGroupProps = PropsWithChildren<{
-    defaultExpanded?: boolean;
-    fieldId: string;
-}>;
-```
