+++ docs/answers-react-components.filters.filtergroupprovider.md
@@ -1,22 +1,22 @@
 <!-- Do not edit this file. It is automatically generated by API Documenter. -->
 
-[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FilterGroup](./answers-react-components.filters.filtergroup.md)
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FilterGroupProvider](./answers-react-components.filters.filtergroupprovider.md)
 
-## Filters.FilterGroup() function
+## Filters.FilterGroupProvider() function
 
-The Filters.FilterGroup component represents a group of filters to support for searching and collapsing.
+The Filters.FilterGroupProvider component represents a group of filters to support for searching and collapsing.
 
 <b>Signature:</b>
 
 ```typescript
-export declare function FilterGroup(props: FilterGroupProps): JSX.Element;
+export declare function FilterGroupProvider(props: FilterGroupProviderProps): JSX.Element;
 ```
 
 ## Parameters
 
 |  Parameter | Type | Description |
 |  --- | --- | --- |
-|  props | FilterGroupProps | [Filters.FilterGroupProps](./answers-react-components.filters.filtergroupprops.md) |
+|  props | FilterGroupProviderProps | [Filters.FilterGroupProviderProps](./answers-react-components.filters.filtergroupproviderprops.md) |
 
 <b>Returns:</b>
