+++ docs/answers-react-components.filters.filtergroupproviderprops.md
@@ -1,0 +1,16 @@
+<!-- Do not edit this file. It is automatically generated by API Documenter. -->
+
+[Home](./index.md) &gt; [@yext/answers-react-components](./answers-react-components.md) &gt; [Filters](./answers-react-components.filters.md) &gt; [FilterGroupProviderProps](./answers-react-components.filters.filtergroupproviderprops.md)
+
+## Filters.FilterGroupProviderProps type
+
+Props for a [Filters.FilterGroupProvider()](./answers-react-components.filters.filtergroupprovider.md)<!-- -->.
+
+<b>Signature:</b>
+
+```typescript
+export declare type FilterGroupProviderProps = PropsWithChildren<{
+    defaultExpanded?: boolean;
+    fieldId: string;
+}>;
+```
