+++ src/components/Filters/index.ts
@@ -1,11 +1,11 @@
 export { CheckboxOption, CheckboxOptionProps, CheckboxCssClasses, FilterOptionConfig } from './CheckboxOption';
 export { CollapsibleLabel, CollapsibleLabelProps } from './CollapsibleLabel';
 export { CollapsibleSection, CollapsibleSectionProps } from './CollapsibleSection';
-export { Facets, FacetsProps } from './Facets';
+export { FacetsProvider, FacetsProviderProps } from './FacetsProvider';
 export { FiltersContext, useFiltersContext, FiltersContextType } from './FiltersContext';
-export { FilterGroup, FilterGroupProps } from './FilterGroup';
+export { FilterGroupProvider, FilterGroupProviderProps } from './FilterGroupProvider';
 export { FilterGroupContext, useFilterGroupContext, FilterGroupContextType } from './FilterGroupContext';
 export { SearchInput, SearchInputProps } from './SearchInput';
-// export { StaticFilters, StaticFiltersProps } from './StaticFilters';
+// export { StaticFiltersProvider, StaticFiltersProviderProps } from './StaticFiltersProvider';
 export { HierarchicalFacet, HierarchicalFacetProps, HierarchicalFacetCssClasses } from './HierarchicalFacet';
 export { RangeInput, RangeInputProps, RangeInputCssClasses } from './RangeInput';
