+++ tests/__compound-components__/Filters/HierarchicalFacet.tsx
@@ -1,11 +1,10 @@
 import { Fragment } from 'react';
 import { Filters } from '../../../src/components';
-import { Facets } from '../../../src/components/Filters/Facets';
 
 export function HierarchicalFacets(args: Filters.HierarchicalFacetProps): JSX.Element {
   const hierarchicalFacetFieldIds = ['hier'];
   return (
-    <Facets searchOnChange={true}>
+    <Filters.FacetsProvider searchOnChange={true}>
       {facets => {
         const filteredFacets = facets.filter(f => f.options.length > 0);
         return (
@@ -23,6 +22,6 @@ export function HierarchicalFacets(args: Filters.HierarchicalFacetProps): JSX.El
           </>
         );
       }}
-    </Facets>
+    </Filters.FacetsProvider>
   );
 }
